Week 3: FiLM-conditioned diffusion action head, validated end-to-end
Adds DiffusionActionHead (a simplified single-resolution version of the
Diffusion Policy paper's temporal U-Net -- a stack of FiLM-conditioned 1D
residual conv blocks; full multi-resolution down/upsampling isn't earning
its complexity at our short chunk_size) and DiffusionPolicy, which reuses
the Week 2 CNN/proprio encoder as conditioning. Training uses diffusers'
DDPMScheduler for the epsilon-prediction loss plus EMA weights; inference
uses DDIMScheduler for fast sampling -- the same train-DDPM/sample-DDIM
split the reference implementation uses.

Single-task overfit sanity check (30 CPU epochs): masked training loss
0.242 -> 0.067, and DDIM-sampled action chunks on held-out validation obs
got 3x closer to ground truth (ddim_sample_mse 0.372 -> 0.127) -- the
actual sampling process is learning the task, not just the training loss.

Also fixes a real bug this surfaced: LiberoChunkDataset's action
normalization silently upcast to float64 (numpy promotes float32 - float64
min/max to float64), which train_bc.py never caught because it only fed
that tensor into a loss computation, but train_diffusion.py feeds noised
actions directly into a Conv1d and crashed on the dtype mismatch. Action
stats are now forced to float32 at the source.

Refactors masked_mse and glob-resolution out of train_bc.py into shared
training/losses.py and training/data_utils.py now that train_diffusion.py
needs the same logic.

+++ src/vla_diffusion/data/libero_dataset.py
@@ -45,8 +45,8 @@ class LiberoChunkDataset(Dataset):
 
         if action_stats is None:
             all_actions = np.concatenate(all_actions, axis=0)
-            self.action_min = all_actions.min(axis=0)
-            self.action_max = all_actions.max(axis=0)
+            self.action_min = all_actions.min(axis=0).astype(np.float32)
+            self.action_max = all_actions.max(axis=0).astype(np.float32)
         else:
             self.action_min = np.array(action_stats["min"], dtype=np.float32)
             self.action_max = np.array(action_stats["max"], dtype=np.float32)
@@ -64,11 +64,11 @@ class LiberoChunkDataset(Dataset):
     def normalize_action(self, a):
         rng = self.action_max - self.action_min
         rng = np.where(rng < 1e-6, 1.0, rng)
-        return 2 * (a - self.action_min) / rng - 1
+        return (2 * (a - self.action_min) / rng - 1).astype(np.float32)
 
     def denormalize_action(self, a):
         rng = self.action_max - self.action_min
-        return (a + 1) / 2 * rng + self.action_min
+        return ((a + 1) / 2 * rng + self.action_min).astype(np.float32)
 
     def action_stats_dict(self):
         return {"min": self.action_min.tolist(), "max": self.action_max.tolist()}
